Split leaf, branch, and extension node files
