+++ firewood/Cargo.toml
@@ -34,6 +34,7 @@ thiserror = "1.0.38"
 tokio = { version = "1.21.1", features = ["rt", "sync", "macros", "rt-multi-thread"] }
 typed-builder = "0.18.0"
 bincode = "1.3.3"
+bitflags = "2.4.1"
 
 [dev-dependencies]
 criterion = {version = "0.5.1", features = ["async_tokio"]}
