+++ firewood/src/merkle/node.rs
@@ -3,11 +3,12 @@
 
 use crate::shale::{disk_address::DiskAddress, CachedStore, ShaleError, ShaleStore, Storable};
 use bincode::{Error, Options};
+use bitflags::bitflags;
 use enum_as_inner::EnumAsInner;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use sha3::{Digest, Keccak256};
 use std::{
-    fmt::{self, Debug},
+    fmt::Debug,
     io::{Cursor, Read, Write},
     sync::{
         atomic::{AtomicBool, Ordering},
@@ -15,15 +16,28 @@ use std::{
     },
 };
 
+mod branch;
+mod extension;
+mod leaf;
+mod partial_path;
+
+pub use branch::{BranchNode, MAX_CHILDREN, SIZE as BRANCH_NODE_SIZE};
+pub use extension::{ExtNode, SIZE as EXTENSION_NODE_SIZE};
+pub use leaf::LeafNode;
+pub use partial_path::PartialPath;
+
 use crate::merkle::to_nibble_array;
 use crate::nibbles::Nibbles;
 
-use super::{from_nibbles, PartialPath, TrieHash, TRIE_HASH_LEN};
+use super::{from_nibbles, TrieHash, TRIE_HASH_LEN};
 
-pub const MAX_CHILDREN: usize = 16;
-
-const EXT_NODE_SIZE: usize = 2;
-const BRANCH_NODE_SIZE: usize = 17;
+bitflags! {
+    // should only ever be the size of a nibble
+    struct Flags: u8 {
+        const TERMINAL = 0b0010;
+        const ODD_LEN  = 0b0001;
+    }
+}
 
 #[derive(Debug, PartialEq, Eq, Clone)]
 pub struct Data(pub(super) Vec<u8>);
@@ -35,8 +49,14 @@ impl std::ops::Deref for Data {
     }
 }
 
+impl From<Vec<u8>> for Data {
+    fn from(v: Vec<u8>) -> Self {
+        Self(v)
+    }
+}
+
 #[derive(Serialize, Deserialize, Debug)]
-pub(crate) enum Encoded<T> {
+enum Encoded<T> {
     Raw(T),
     Data(T),
 }
@@ -57,292 +77,6 @@ impl<T: DeserializeOwned + AsRef<[u8]>> Encoded<T> {
     }
 }
 
-#[derive(PartialEq, Eq, Clone)]
-pub struct BranchNode {
-    pub(super) path: PartialPath,
-    pub(super) children: [Option<DiskAddress>; MAX_CHILDREN],
-    pub(super) value: Option<Data>,
-    pub(super) children_encoded: [Option<Vec<u8>>; MAX_CHILDREN],
-}
-
-impl Debug for BranchNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f, "[Branch")?;
-        write!(f, " path={:?}", self.path)?;
-
-        for (i, c) in self.children.iter().enumerate() {
-            if let Some(c) = c {
-                write!(f, " ({i:x} {c:?})")?;
-            }
-        }
-
-        for (i, c) in self.children_encoded.iter().enumerate() {
-            if let Some(c) = c {
-                write!(f, " ({i:x} {:?})", c)?;
-            }
-        }
-
-        write!(
-            f,
-            " v={}]",
-            match &self.value {
-                Some(v) => hex::encode(&**v),
-                None => "nil".to_string(),
-            }
-        )
-    }
-}
-
-impl BranchNode {
-    pub(super) fn single_child(&self) -> (Option<(DiskAddress, u8)>, bool) {
-        let mut has_chd = false;
-        let mut only_chd = None;
-        for (i, c) in self.children.iter().enumerate() {
-            if c.is_some() {
-                has_chd = true;
-                if only_chd.is_some() {
-                    only_chd = None;
-                    break;
-                }
-                only_chd = (*c).map(|e| (e, i as u8))
-            }
-        }
-        (only_chd, has_chd)
-    }
-
-    pub fn decode(buf: &[u8]) -> Result<Self, Error> {
-        let mut items: Vec<Encoded<Vec<u8>>> = bincode::DefaultOptions::new().deserialize(buf)?;
-
-        // we've already validated the size, that's why we can safely unwrap
-        let data = items.pop().unwrap().decode()?;
-        // Extract the value of the branch node and set to None if it's an empty Vec
-        let value = Some(data).filter(|data| !data.is_empty());
-
-        // encode all children.
-        let mut chd_encoded: [Option<Vec<u8>>; MAX_CHILDREN] = Default::default();
-
-        // we popped the last element, so their should only be NBRANCH items left
-        for (i, chd) in items.into_iter().enumerate() {
-            let data = chd.decode()?;
-            chd_encoded[i] = Some(data).filter(|data| !data.is_empty());
-        }
-
-        // TODO: add path
-        let path = Vec::new().into();
-
-        Ok(BranchNode::new(
-            path,
-            [None; MAX_CHILDREN],
-            value,
-            chd_encoded,
-        ))
-    }
-
-    fn encode<S: ShaleStore<Node>>(&self, store: &S) -> Vec<u8> {
-        // TODO: add path to encoded node
-        let mut list = <[Encoded<Vec<u8>>; MAX_CHILDREN + 1]>::default();
-
-        for (i, c) in self.children.iter().enumerate() {
-            match c {
-                Some(c) => {
-                    let mut c_ref = store.get_item(*c).unwrap();
-
-                    if c_ref.is_encoded_longer_than_hash_len::<S>(store) {
-                        list[i] = Encoded::Data(
-                            bincode::DefaultOptions::new()
-                                .serialize(&&(*c_ref.get_root_hash::<S>(store))[..])
-                                .unwrap(),
-                        );
-
-                        // See struct docs for ordering requirements
-                        if c_ref.lazy_dirty.load(Ordering::Relaxed) {
-                            c_ref.write(|_| {}).unwrap();
-                            c_ref.lazy_dirty.store(false, Ordering::Relaxed)
-                        }
-                    } else {
-                        let child_encoded = &c_ref.get_encoded::<S>(store);
-                        list[i] = Encoded::Raw(child_encoded.to_vec());
-                    }
-                }
-                None => {
-                    // Check if there is already a calculated encoded value for the child, which
-                    // can happen when manually constructing a trie from proof.
-                    if let Some(v) = &self.children_encoded[i] {
-                        if v.len() == TRIE_HASH_LEN {
-                            list[i] =
-                                Encoded::Data(bincode::DefaultOptions::new().serialize(v).unwrap());
-                        } else {
-                            list[i] = Encoded::Raw(v.clone());
-                        }
-                    }
-                }
-            };
-        }
-
-        if let Some(Data(val)) = &self.value {
-            list[MAX_CHILDREN] =
-                Encoded::Data(bincode::DefaultOptions::new().serialize(val).unwrap());
-        }
-
-        bincode::DefaultOptions::new()
-            .serialize(list.as_slice())
-            .unwrap()
-    }
-
-    pub fn new(
-        path: PartialPath,
-        chd: [Option<DiskAddress>; MAX_CHILDREN],
-        value: Option<Vec<u8>>,
-        chd_encoded: [Option<Vec<u8>>; MAX_CHILDREN],
-    ) -> Self {
-        BranchNode {
-            path,
-            children: chd,
-            value: value.map(Data),
-            children_encoded: chd_encoded,
-        }
-    }
-
-    pub fn value(&self) -> &Option<Data> {
-        &self.value
-    }
-
-    pub fn chd(&self) -> &[Option<DiskAddress>; MAX_CHILDREN] {
-        &self.children
-    }
-
-    pub fn chd_mut(&mut self) -> &mut [Option<DiskAddress>; MAX_CHILDREN] {
-        &mut self.children
-    }
-
-    pub fn chd_encode(&self) -> &[Option<Vec<u8>>; MAX_CHILDREN] {
-        &self.children_encoded
-    }
-
-    pub fn chd_encoded_mut(&mut self) -> &mut [Option<Vec<u8>>; MAX_CHILDREN] {
-        &mut self.children_encoded
-    }
-}
-
-#[derive(PartialEq, Eq, Clone)]
-pub struct LeafNode {
-    pub(super) path: PartialPath,
-    pub(super) data: Data,
-}
-
-impl Debug for LeafNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        write!(f, "[Leaf {:?} {}]", self.path, hex::encode(&*self.data))
-    }
-}
-
-impl LeafNode {
-    fn encode(&self) -> Vec<u8> {
-        bincode::DefaultOptions::new()
-            .serialize(
-                [
-                    Encoded::Raw(from_nibbles(&self.path.encode(true)).collect()),
-                    Encoded::Raw(self.data.to_vec()),
-                ]
-                .as_slice(),
-            )
-            .unwrap()
-    }
-
-    pub fn new(path: Vec<u8>, data: Vec<u8>) -> Self {
-        LeafNode {
-            path: PartialPath(path),
-            data: Data(data),
-        }
-    }
-
-    pub fn path(&self) -> &PartialPath {
-        &self.path
-    }
-
-    pub fn data(&self) -> &Data {
-        &self.data
-    }
-}
-
-#[derive(PartialEq, Eq, Clone)]
-pub struct ExtNode {
-    pub(crate) path: PartialPath,
-    pub(crate) child: DiskAddress,
-    pub(crate) child_encoded: Option<Vec<u8>>,
-}
-
-impl Debug for ExtNode {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
-        let Self {
-            path,
-            child,
-            child_encoded,
-        } = self;
-        write!(f, "[Extension {path:?} {child:?} {child_encoded:?}]",)
-    }
-}
-
-impl ExtNode {
-    fn encode<S: ShaleStore<Node>>(&self, store: &S) -> Vec<u8> {
-        let mut list = <[Encoded<Vec<u8>>; 2]>::default();
-        list[0] = Encoded::Data(
-            bincode::DefaultOptions::new()
-                .serialize(&from_nibbles(&self.path.encode(false)).collect::<Vec<_>>())
-                .unwrap(),
-        );
-
-        if !self.child.is_null() {
-            let mut r = store.get_item(self.child).unwrap();
-
-            if r.is_encoded_longer_than_hash_len(store) {
-                list[1] = Encoded::Data(
-                    bincode::DefaultOptions::new()
-                        .serialize(&&(*r.get_root_hash(store))[..])
-                        .unwrap(),
-                );
-
-                if r.lazy_dirty.load(Ordering::Relaxed) {
-                    r.write(|_| {}).unwrap();
-                    r.lazy_dirty.store(false, Ordering::Relaxed);
-                }
-            } else {
-                list[1] = Encoded::Raw(r.get_encoded(store).to_vec());
-            }
-        } else {
-            // Check if there is already a caclucated encoded value for the child, which
-            // can happen when manually constructing a trie from proof.
-            if let Some(v) = &self.child_encoded {
-                if v.len() == TRIE_HASH_LEN {
-                    list[1] = Encoded::Data(bincode::DefaultOptions::new().serialize(v).unwrap());
-                } else {
-                    list[1] = Encoded::Raw(v.clone());
-                }
-            }
-        }
-
-        bincode::DefaultOptions::new()
-            .serialize(list.as_slice())
-            .unwrap()
-    }
-
-    pub fn chd(&self) -> DiskAddress {
-        self.child
-    }
-
-    pub fn chd_encoded(&self) -> Option<&[u8]> {
-        self.child_encoded.as_deref()
-    }
-
-    pub fn chd_mut(&mut self) -> &mut DiskAddress {
-        &mut self.child
-    }
-
-    pub fn chd_encoded_mut(&mut self) -> &mut Option<Vec<u8>> {
-        &mut self.child_encoded
-    }
-}
-
 #[derive(Debug)]
 pub struct Node {
     pub(super) root_hash: OnceLock<TrieHash>,
@@ -356,6 +90,7 @@ pub struct Node {
 }
 
 impl Eq for Node {}
+
 impl PartialEq for Node {
     fn eq(&self, other: &Self) -> bool {
         let Node {
@@ -372,6 +107,7 @@ impl PartialEq for Node {
             && *inner == other.inner
     }
 }
+
 impl Clone for Node {
     fn clone(&self) -> Self {
         Self {
@@ -397,7 +133,7 @@ impl NodeType {
         let items: Vec<Encoded<Vec<u8>>> = bincode::DefaultOptions::new().deserialize(buf)?;
 
         match items.len() {
-            EXT_NODE_SIZE => {
+            EXTENSION_NODE_SIZE => {
                 let mut items = items.into_iter();
 
                 let decoded_key: Vec<u8> = items.next().unwrap().decode()?;
@@ -405,7 +141,8 @@ impl NodeType {
                 let decoded_key_nibbles = Nibbles::<0>::new(&decoded_key);
 
                 let (cur_key_path, term) =
-                    PartialPath::from_nibbles(decoded_key_nibbles.into_iter());
+                    dbg!(PartialPath::from_nibbles(decoded_key_nibbles.into_iter()));
+
                 let cur_key = cur_key_path.into_inner();
                 let data: Vec<u8> = items.next().unwrap().decode()?;
 
@@ -550,28 +287,33 @@ impl Node {
 
 impl Storable for Node {
     fn deserialize<T: CachedStore>(addr: usize, mem: &T) -> Result<Self, ShaleError> {
-        const META_SIZE: usize = 32 + 1 + 1;
+        const META_SIZE: usize = TRIE_HASH_LEN + 1 + 1;
+
         let meta_raw =
             mem.get_view(addr, META_SIZE as u64)
                 .ok_or(ShaleError::InvalidCacheView {
                     offset: addr,
                     size: META_SIZE as u64,
                 })?;
-        let attrs = meta_raw.as_deref()[32];
+
+        let attrs = meta_raw.as_deref()[TRIE_HASH_LEN];
+
         let root_hash = if attrs & Node::ROOT_HASH_VALID_BIT == 0 {
             None
         } else {
             Some(TrieHash(
-                meta_raw.as_deref()[0..32]
+                meta_raw.as_deref()[..TRIE_HASH_LEN]
                     .try_into()
                     .expect("invalid slice"),
             ))
         };
+
         let is_encoded_longer_than_hash_len = if attrs & Node::IS_ENCODED_BIG_VALID == 0 {
             None
         } else {
             Some(attrs & Node::LONG_BIT != 0)
         };
+
         match meta_raw.as_deref()[33] {
             Self::BRANCH_NODE => {
                 // TODO: add path
@@ -650,6 +392,7 @@ impl Storable for Node {
                     }),
                 ))
             }
+
             Self::EXT_NODE => {
                 let ext_header_size = 1 + 8;
                 let node_raw = mem.get_view(addr + META_SIZE, ext_header_size).ok_or(
@@ -688,9 +431,12 @@ impl Storable for Node {
                         offset: addr + META_SIZE + ext_header_size as usize + path_len as usize,
                         size: 1,
                     })?;
+
                 cur = Cursor::new(encoded_len_raw.as_deref());
                 cur.read_exact(&mut buff)?;
+
                 let encoded_len = buff[0] as u64;
+
                 let encoded: Option<Vec<u8>> = if encoded_len != 0 {
                     let emcoded_raw = mem
                         .get_view(
@@ -711,7 +457,7 @@ impl Storable for Node {
                     None
                 };
 
-                Ok(Self::new_from_hash(
+                let node = Self::new_from_hash(
                     root_hash,
                     is_encoded_longer_than_hash_len,
                     NodeType::Extension(ExtNode {
@@ -719,8 +465,11 @@ impl Storable for Node {
                         child: DiskAddress::from(ptr as usize),
                         child_encoded: encoded,
                     }),
-                ))
+                );
+
+                Ok(node)
             }
+
             Self::LEAF_NODE => {
                 let leaf_header_size = 1 + 4;
                 let node_raw = mem.get_view(addr + META_SIZE, leaf_header_size).ok_or(
@@ -729,11 +478,15 @@ impl Storable for Node {
                         size: leaf_header_size,
                     },
                 )?;
+
                 let mut cur = Cursor::new(node_raw.as_deref());
                 let mut buff = [0; 4];
                 cur.read_exact(&mut buff[..1])?;
+
                 let path_len = buff[0] as u64;
+
                 cur.read_exact(&mut buff)?;
+
                 let data_len = u32::from_le_bytes(buff) as u64;
                 let remainder = mem
                     .get_view(
@@ -753,12 +506,15 @@ impl Storable for Node {
                     .collect();
 
                 let (path, _) = PartialPath::decode(&nibbles);
-                let value = Data(remainder.as_deref()[path_len as usize..].to_vec());
-                Ok(Self::new_from_hash(
+                let data = Data(remainder.as_deref()[path_len as usize..].to_vec());
+
+                let node = Self::new_from_hash(
                     root_hash,
                     is_encoded_longer_than_hash_len,
-                    NodeType::Leaf(LeafNode { path, data: value }),
-                ))
+                    NodeType::Leaf(LeafNode { path, data }),
+                );
+
+                Ok(node)
             }
             _ => Err(ShaleError::InvalidNodeType),
         }
@@ -787,13 +543,13 @@ impl Storable for Node {
                 }
                 NodeType::Extension(n) => {
                     1 + 8
-                        + n.path.dehydrated_len()
+                        + n.path.serialized_len()
                         + match n.chd_encoded() {
                             Some(v) => 1 + v.len() as u64,
                             None => 1,
                         }
                 }
-                NodeType::Leaf(n) => 1 + 4 + n.path.dehydrated_len() + n.data.len() as u64,
+                NodeType::Leaf(n) => 1 + 4 + n.path.serialized_len() + n.data.len() as u64,
             }
     }
